Completed most necessary functions.
Submission function handling each case and auto saving code for each parts as well as progressing the main file to the next part if necessary. Create day function to run all the necessary functions to create a folder containing a main script, a test input file, and the main input file fetched from the AoC website.